refactor: clean up hook imports
