+++ src/features/transfer-ownership/TransferOwnershipForm/hooks/useTransferOwnershipForm.tsx
@@ -7,9 +7,7 @@ import {
 
 import { Step } from '../FormSteps/hooks';
 import { getDomainId } from '../../../../lib/util';
-import { useWeb3 } from '../../../../lib/hooks/useWeb3';
-import { useZnsSdk } from '../../../../lib/hooks/useZnsSdk';
-import { useDomainData } from '../../../../lib/hooks/useDomainData';
+import { useWeb3, useZnsSdk, useDomainData } from '../../../../lib/hooks';
 import { useTransaction } from '@zero-tech/zapp-utils/hooks/useTransaction';
 
 enum StatusText {
