+++ src/features/view-nft/NFTDetailsCard/NFTDetailsCard.tsx
@@ -2,10 +2,8 @@ import { FC } from 'react';
 
 import { MemberTitle } from '../../../lib/constants/labels';
 import { getDomainId } from '../../../lib/util/domains/domains';
-import { useDomainData } from '../../../lib/hooks/useDomainData';
-import { useDomainMetadata } from '../../../lib/hooks/useDomainMetadata';
+import { useDomainData, useDomainMetadata } from '../../../lib/hooks';
 
-import { CTAContainer } from './ui';
 import { Options } from './ui/Options';
 import { Member } from '../../ui/Member';
 import { SkeletonText, SkeletonTextProps } from '@zero-tech/zui/components';
